Log 708231
Problem: When called from mai3820, pressing the defect button on the summary screen does not move focus
to the detail window. 

Solution: Problem occured because the the defect button was always pointing to the b1.s_road_id
on the summary screen. Modified the when-window-activated trigger to move to the appropriate field on the defect screen 
and also to improve performance issues when moving between the detail and summary screens - further details provided in the trigger.

git-svn-id: svn://127.0.0.1/mai@99 521c773d-b55a-d943-8648-67ffa5d26747