Log 708835
Problem: When focus is on the authorise button and you try to move to another screen the message 'Function cannot be performed here' is displayed. If the cursor is in the WO number field a prompt appears asking you if you want to save before going into query mode. The expand button on the toolbar is disabled so you cannot drill down to mai3800 to access details of the works order.

Solution: Commented out the do_key('enter_query') in the when-window activated trigger. Enabled the expand
button and ensured it called mai3800.

git-svn-id: svn://127.0.0.1/mai@102 521c773d-b55a-d943-8648-67ffa5d26747