log: 708476 problem: form is missing lov and button on b0.WPR_ID and missing lov button on b0.WPR_INT_CODE.
solution: The lov for b0.WPR_ID already existed, so I assigned it to the field and added an LOV button.
Edited the code in the LOV B0_WPR_ID_LOV1 to only return Priorities from the hig_code table where the start/end dates are compliant with the system effective date.
added a button for b0.WPR_INT_CODE to call the LOV.

git-svn-id: svn://127.0.0.1/mai@142 521c773d-b55a-d943-8648-67ffa5d26747