Log 709407 Problem: DRD logged changes to mai3863, they wanted to specify a default PED filename and restirct the list of inspectors in the PED file by the users admin unit. Solution: New Variable: Called FAunit, which is autopopulated with the gri data from the parameter ANSWER2, Y/N restrict the inspectors by the users admin uit. The variable FFile willalso be populated with the gri data from the parameter PEDFILE, which is the default filename for the PED file.
dbmsoutput:
additional output lines added for Output file and restricted AU

New Cursor:
C31a
Same as C31 which returns all the inspectors, but with additional where clause which matches the AU of the User to the AU of the admin group.  Hierarchically

Additional code:
Call data from cursors C31 or C31a depending on wether the variable FAunit is equal to 'N' (C31) or not(C31a).

Product options:
There will be a new user option called PEDFILE which will allow the user to default the filename when a PED is created.

Related changes:
maidata1.sql        delta 2.3
maidata2.sql	delta 2.3

git-svn-id: svn://127.0.0.1/mai@232 521c773d-b55a-d943-8648-67ffa5d26747