log: 710397 Problem: Dorset logged issue when viewing inspection due dates calculated on a frequency, the dates aren't calculated correctly.
Solution: My fix from october last year was incorrect.  I should not have updated the queries in the report, but should have altered the dynamic SQL in the before report trigger which inserts data into the SCHEDULE and SCHEDULE_TEMP tables which the report queries are based on.
So i have removed all my changes from the queries and instead altered the dynamic SQL to populate the SCHEDULE_TEMP table with ROAD_SECTIONS.RSE_LAST_INSPECTED rather than ACTIVITIES_REPORT.ARE_DATE_WORK_DONE field.   This means that DATE_DUE and the STATUS fields should now be correct as they are based on the date the last inspection was done.

git-svn-id: svn://127.0.0.1/mai@255 521c773d-b55a-d943-8648-67ffa5d26747