Log 713476
Problem: Cannot add multiple defects to inspections.

Solution: Problem occurs when an additional defect is created with a different activity code. When the repair details are entered the rep_atv_acty_area_code item should be assigned the defect's activity code via the initial value property. This works ok in 6i but not 9i/10g. To fix, I've mannally assigned the value to the rep_atv_acty_area_code item in the when-new-block-instance trigger on the repairs block.

git-svn-id: svn://127.0.0.1/mai@488 521c773d-b55a-d943-8648-67ffa5d26747