log 709821,704064 problem:on opening the form in enter-query mode the message, FRM-41361 cannot navigate out of form in enter-query mode is shown.  After the form has been closed, the message FRM-41351 is shown.
solution: moved the  enter_query from form_startup to when-window_activated to avoid the FRM-41361/FRM-41351 errors when entering/leaving the form.  This was caused by the form automatically being in enter_query mode when the form was opened.  A check is done to only place the form into enter_query mode when the form is first opened.

git-svn-id: svn://127.0.0.1/mai@526 521c773d-b55a-d943-8648-67ffa5d26747