Log 715508
Problem: 1) A non-oracle exception message occurs when entering the form. 2) Closing the inventory item detail window displays a frm-41032 message  which is caused by attempting to enable the inventory item button whist the focus was still on it.

Fix: Modified the post-query trigger on B9 by commenting out the call to plib$fail procedure which caused the non-oracle exception. Changed the when-new-block-instance trigger on B7 to move to the next item if the focus was still on the inventory detail button.

git-svn-id: svn://127.0.0.1/mai@544 521c773d-b55a-d943-8648-67ffa5d26747