716119
Added code to store the state of the defects before mai3801 updates them so that it can restore any defects that don't make it on to the work order.

Also removed code that updates the budgets as the work order created is not instructed and mai no-longer updates budgets until the work order has been instructed.

git-svn-id: svn://127.0.0.1/mai@586 521c773d-b55a-d943-8648-67ffa5d26747