The following issues raised by HB:-
"The Form needs to be able to display minimum 6 characters in the �Items� and �Item Errors�, now that the limit of 9999 records per file no longer exists."
These fields have been expanded to 10 digits and made wider on the camvas.

"A �D� seems to be forced into the Sys Flag field when modifications are made to data."
The form has been changed to get the sys flag from road_segs rather than making assumptions based upon the length of the Section Number.

git-svn-id: svn://127.0.0.1/mai@610 521c773d-b55a-d943-8648-67ffa5d26747