Log 719274
Problem - the fact that the rse_descr has been changed in MAI3806 and MAI3800 from 80 to 240 chars is causing a problem when changing the work order status in MAI3800 as it is now possible to try and insert a rse_descr value > 80 chars into the interface_wol table which is causing it to fall over.

Solution - restricted the rse_descr to 80 chars as per GNA's instructions using SUBSTR.  Anything higher is more than interface_wol can handle.  This is a temporary measure until an impact anlysis of increasing the rse_descr from 80 to 240 is undertaken and the changes to the relevant forms/reports made. 

git-svn-id: svn://127.0.0.1/mai@744 521c773d-b55a-d943-8648-67ffa5d26747