Log 721215 Problem - when splitting a network, the following error is given: ORA-20000: HIG-0436: User is not permitted to operate on the selected network element ORA-06512: at 'HIGHWAYS.HIG', line 1476 A user with the highest admin level is editing a maintenance network but the error is returned
Soultion - made changes as per Ade's suggestion to procedure create_new_defect.
The cursor which derives the das_table_name from doc_assocs for the old defect is bringing back all sorts of doc_assocs � in this case an Element which has the same primary key as the old defect. 

The code needs to restrict on either DEFECTS or a Doc gateway synonym of DEFECTS when trying to find associated doc_assocs � and the call to NM3RECLASS.INS_DOC_ASSOCS needs to deal with no data returned from the cursor, as it will currently fail when no doc assoc are found.


git-svn-id: svn://127.0.0.1/mai@947 521c773d-b55a-d943-8648-67ffa5d26747