Added for Graeme - Mistakenly put in f60
form and it'll enter query mode. Why.......�

Well I opened the form and found that most of the forms level triggers which should be subclassed were not actually subclassed and that someone had been left to code around this and ended up putting a do_key(�enter_query�) call on the when-window-activated trigger.

It makes me weep

git-svn-id: svn://127.0.0.1/mai@1083 521c773d-b55a-d943-8648-67ffa5d26747