It appears to be because the value of product option WORSTDTXT is 2000 characters.  It�s trying to put that into an 80 character variable.
I know that at release 4.1.0.0 the size of product option values  has been increased up to 2000 characters.

To change the size of the �standard text� on the layout model would mean having to play about with the whole report layout.  

This is because it sits slap bang in the middle of the report and to make it expandable has a knock on effect

My �solution�  has been just to change the size of the placeholder only.

git-svn-id: svn://127.0.0.1/mai@1111 521c773d-b55a-d943-8648-67ffa5d26747