Two issues resolved...
Task  0108286
============
Rationalised the WNRI trigger on the DEF block which enables/disables, sets items for update/non-update etc etc.

There was some logic regarding setting of the update permission of defect priority i.e.

     IF product option DEFAUTOPRI  = 'A' THEN
 
                                 Disable the Defect Priority field
 
     ELSE IF product option DEFAUTOPRI  = 'B' THEN

                                 Enable the Defect Priority field
                                 The Defect Priority is optional

     ELSE

                                 Enable the Defect Priority field
                                 The Defect Priority is Mandatory

     END IF;


Amended this to...


     IF defect is NOT Available (i.e. associated feature code 5 = `N�) OR  product option DEFAUTOPRI  = 'A' THEN

                                 Disable the Defect Priority field


     ELSE IF product option DEFAUTOPRI  = 'B' THEN

                                 Enable the Defect Priority field
                                 The Defect Priority is optional

     ELSE

                                 Enable the Defect Priority field
                                 The Defect Priority is Mandatory

     END IF;



Task 0108564
===========
Amended the form so the 'Defects' button is now only enabled when
a) there is an inspection id (top left hand corner of the form)
AND

b) the road section attributed to the inspection is NOT end dated
OR 
b) there are existing defects attributed to the inspection


git-svn-id: svn://127.0.0.1/mai@1161 521c773d-b55a-d943-8648-67ffa5d26747