Task 0107322 Fix the WVR trigger that was asking the user to save changes when they were pressing the commit button, also an error that was being raised when changing the status and pressing the boq button without first tabbing out of the status item.
Task 0107328
program unit defect_status now only returns a value when the specified conditions are met i.e. the defect status needs to be changed.
The program unit update_defect_status has been changed to only do the update if the status passed in is not null.

Task 0108554
Changed WVR trigger on B1 to not check that wor_descr is populated before asking the user to commit.

Task 0108570
:b0.h_dumconcode expanded to 2000 characters

Task 0107444
Re-writen the code that deals with changing contract via the contract totals screen, now uses the same mechanism to recalculate the Work Order costs as a direct change of contract in the Work Order Header (main canvas).

Task 0107354
Additional fields for road description and defect description added to the Defect Selection canvas.
Also added dynamic ordering to the block and altered the code that filters available defects (select_defects2) to deal with more than 1000 defects.

Task 0107298
Corrected a discrepancy between the Standard and Detailed BOQ LOVs (the Detailed one is used when Product Option STADETAIL is set to either 'Y').
Previously the detailed LOV was allowing the selection of enddated Standard Items when a  Work Order was assigned to the DUMMY contract, these then fail validation and an error is diaplayed to the user.
Also tweaked both LOVs and validation to prevent the selection of Percentage Items as "Parent" Items when the contract is not the dummy contract.

git-svn-id: svn://127.0.0.1/mai@1206 521c773d-b55a-d943-8648-67ffa5d26747