Task 0108773 Altered Percentage Items LOV to restrict by contract start and end date.
Task 0108778
Form updated to reprice percentage items when the contract is changed.
Also corrected the calculation of the "% Order Lines" field on the contract totals canvas as it was showing more than 100% when all the boqs are on the contract and there are also percentage items on the lines.
This calculation excludes percentage items so if all the parent BOQs are on the contract but a percentage item isn't the form will display 100%. The question of whether or not percentage items should be included in this calculation will be dealt with in another task/release.

git-svn-id: svn://127.0.0.1/mai@1235 521c773d-b55a-d943-8648-67ffa5d26747