Task 0109370
Fixed the problem of a null estimated rate when new boqs are added via mai3802, estimated rate is now set to 0 in line with mai3800.

Task 0109311

Altered the query that drives the BOQ block in MAI3800 (Work Orders) and MAI3802 (Maintain Work Orders - Contractor Interface).
This was done to exclude 'phantom' items that appear when a parent item has a percentage item attached which in turn has another percentage item attached (i.e. three levels of boq items)

git-svn-id: svn://127.0.0.1/mai@1349 521c773d-b55a-d943-8648-67ffa5d26747