Task 0109382
MAI3800 (Work Orders) was asking the user to save changes after querying a Work Order Line with a Gang on it.

Task 0109306
Changed so that if WORREFGEN is set to 'A' the Admin Unit of the selected Contract is used to calculate the Work Order Number.
If however the Default (dummy) Contract has been selected OR the new Product Option (WORREFUSER) has been set to 'Y' the Admin Unit of the User that creates the Work Order will be used to generate the Work Order Number.

Task 0109262
MAI3800 (Work Orders) The form was not recalculating the cost of a Work Order on change of contract when the product option WORREFGEN is set to 'A'.

git-svn-id: svn://127.0.0.1/mai@1350 521c773d-b55a-d943-8648-67ffa5d26747