Task 0109505
Following on from version 2.14, relaced all remaining occurences of oracle reserved word USER with the call to nm3context.
This has been done separately so that the minimal change in version 2.14 can be release as part of a fix for 4200, this change will be released as part of 4300 and will require sifgificantly more system testing.

git-svn-id: svn://127.0.0.1/mai@1375 521c773d-b55a-d943-8648-67ffa5d26747