Task 0109946 Allow users to alter the BOQ Item Code on an Instructed WOL if the estimated rate is NULL. This happens when a defect that already has associated BOQs is attached to an instructed WO and a BOQ does not match the contract.
Task 0110619
Instruction of a WO has been changed so that the WOR_DATE_CONFIRMED is not enterable at any time, when the user presses the Instruct button a popup is displayed allowing the user to change the default date instructed (sysdate) before continuing.

Task 0110686
Have made the field enterable to allow the user to navigate to the field and copy the WO Number.
The form still prevents the user from manually entering or changing the WO Number when WORREFGEN is 'A' or 'C' and will allow the user to enter a manual WO Number if WORREFGEN is 'M' and no Works Order Lines have been added to the WO.

git-svn-id: svn://127.0.0.1/mai@2101 521c773d-b55a-d943-8648-67ffa5d26747